requirements.txt
Flask
pywebview
ctransformers
diffusers
torch
accelerate
Pillow